fix: deep-review batch - GPU perf, safety, overflow, docs (#43 #48 #49 #50 #58)
- Remove sRGB render target round-trip: Rgba8UnormSrgb -> Rgba8Unorm,
  drop srgb_to_linear/linear_to_srgb from shaders (~50M pow() calls/frame)
- Remove depth buffer: 2 fixed-order quads don't need depth testing,
  saves 33MB/frame bandwidth at 4K
- Replace unsafe Surface transmute with Arc<Window> (wgpu 25 API)
- Fix field drop order: surface before window to prevent use-after-free
- Add tune=zerolatency for x264, bump CRF for fast preset, 4 threads on Jetson
- Cast buffer sizes to usize/u64 before multiplication to prevent u32 overflow
- Add MAX_DIM=8192 validation on CLI dimension args
- Add bounds checks in FFmpeg extract_plane
- Update all stale docs: reco-ffmpeg -> reco-io, NV12 support, GStreamer exists

Closes #43, closes #48, closes #49, closes #50, closes #58

+++ crates/reco-core/src/nv12_converter.rs
@@ -133,11 +133,11 @@ impl Nv12Converter {
         });
 
         // NV12 buffer size: Y plane (w*h) + UV plane (w*h/2) = w*h*3/2 bytes
-        let nv12_bytes = width * height * 3 / 2;
+        let nv12_bytes = width as u64 * height as u64 * 3 / 2;
         // GPU storage buffer for compute shader output
         let nv12_gpu_buffer = device.create_buffer(&wgpu::BufferDescriptor {
             label: Some("nv12_gpu"),
-            size: nv12_bytes as u64,
+            size: nv12_bytes,
             usage: wgpu::BufferUsages::STORAGE | wgpu::BufferUsages::COPY_SRC,
             mapped_at_creation: false,
         });
@@ -145,7 +145,7 @@ impl Nv12Converter {
         // CPU-readable staging buffer for readback
         let nv12_staging_buffer = device.create_buffer(&wgpu::BufferDescriptor {
             label: Some("nv12_staging"),
-            size: nv12_bytes as u64,
+            size: nv12_bytes,
             usage: wgpu::BufferUsages::COPY_DST | wgpu::BufferUsages::MAP_READ,
             mapped_at_creation: false,
         });
